Make DELETE of nonexistent resource succeed
If we ask to DELETE something and it already doesn't exist, that
should be considered successful since the postcondition of our
request is met.

This is important because it makes DELETE operations idempotent;
without this, a delete interrupted by a network glitch may not
be able to complete by retrying.