tests: fix test_search_stops unstable (on Python 2.x only?)
This test is trying to ensure that requests for further pages
are cancelled if the caller no longer holds a reference to the
page. It was flaky because there was no way to guarantee that
the cancellation (from one thread) would happen before another
thread would start the request for the next page.

Forcing a retry delay after the first couple of pages makes it
stable, since it gives a predictable delay during which the
cancel can happen.

Although it seems like this could have been flaky for both
Python 2 and 3, in practice it was only seen to fail on Python
2, maybe due to differences in GC or threading implementation.