tests: fix flaky test_publish_cancel
This test was trying to verify that a Pulp task was cancelled
via a DELETE request. It had been written to assume that the
DELETE would always be the last request made, but that doesn't
make sense - the task poller thread is still running and could
start a search request after the DELETE.

Let's just verify that the expected DELETE request exists,
but not require it to be the very last request.