Support searching by regex, with minor refactors
We have a use-case for searching repositories by regular expression,
so let's implement that.

Includes some backwards-compatible refactoring of the Criteria API.
Criteria and Matcher are now two separate concepts. This avoids
having to implement e.g. "with_field", "with_field_regex" and so
on methods for any type of matching we might want to do.