Allow searching on model fields rather than (only) raw Pulp fields
Previously, it was necessary always to use Pulp field names in
repo searches.  This was awkward, as it means the caller needs to
know and deal with both the Pulp and the model field names for
the same piece of data.

For example, to find Repository objects with eng_product_id=123,
the caller would have to know that this field is stored as a
string under "notes.eng_product", thus the model wouldn't fully
encapsulate the Pulp implementation details.

Improve this by making it possible to provide the model field
names to searches.  If the caller wants to find repositories
with (for example) eng_product_id=123, they can now directly
search with that field name and value.  This should be
preferred, as the caller no longer has to know exactly how a field
has been stored within Pulp.

However, since it's not usable under all circumstances, search
on raw Pulp fields will remain supported.