Fix inconsistent handling of bad criteria
With the real client, if you'd passed a bad criteria object, you'd
immediately get an exception. On the fake client, you'd rather get
a future resolved with an exception.

It's best to be consistent between the two clients, so let's adjust
the fake client to also raise immediately.  We can share the real
client code to do this.