Introduce Client.get_content_type_ids
Simple API to obtain the content types supported by the Pulp server.

The use-case for this API is the clear-repo command. This command
is able to accept specific content types to operate on. We'll need
to validate content types passed by the user, which means we need
to know which content types are supported by the server.