Added less_than matcher for the criteria
less_than matcher returns the objects whose field values
is less than the matcher value.
Dates can also be compared with this matcher. It requires the dates
to be in a specific ISO format so as to identify it as a date and
generate the criteria accordingly.
Also, the fake client is updated for this new matcher.
Corresponding tests were added as well.