Deprecate Page.as_iter in preference of __iter__
Implementing __iter__ allows to directly iterate over Page objects
rather than having to call as_iter.

The original idea with adding Page.as_iter rather than __iter__ was
that using as_iter is blocking, and it would be better to be
explicit rather than implicit whenever operations might block, so
as to help keep a workflow entirely non-blocking.

However, now we have several pieces of code using this library and
the following pattern is seen commonly:

  repos = client.search_repository(...).result().as_iter()
  for repo in repos:
    ...

That's a bit too noisy, and this usage seems to be more common than
using search results in a non-blocking manner anyway (e.g. we usually
want to find & validate *all* repos before moving to the next step).

So, it seems like the idea of "explicit over implicit" here isn't
helping. Let's allow using the result as an iterable directly, so
above code can be written as slightly cleaner:

  repos = client.search_repository(...).result()
  for repo in repos:
    ...