Fix bogus field conversions in fake client
In the fake client, our logic for deciding whether we had found a
model field or a raw Pulp field was:

  using_model_field = mapped_field is not field

But that clearly doesn't work if the model field and the Pulp field
happen to have the same name.  This would lead to fields wrongly
using a Pulp conversion where none is required.

Up to now, it seems no fields hit the conditions where this mattered,
which is a field having:

- the same name on model and in Pulp
- and a (nontrivial) conversion required

With the addition of Distributor.last_publish as a datetime, this
bug now matters and must be fixed.

Instead of trying to reuse a single piece of data to mean multiple
things, let's just define the function as returning None if no
mapping occurred.