refactor: use FrozenList class for immutable lists
The intent is for all public model classes to be immutable.
However, a problem arises when a model is expected to have a list
field, as lists are mutable and not hashable.

Previously, the attempted fix for this was to use tuples for those
fields instead.  This was done since a tuple is the closest thing
to an immutable list built-in to Python.

Unfortunately it turns out there are some undesirable properties
from using tuples.  The biggest issue is that most APIs are defined
in terms of lists and developers are usually picking lists as the
default choice for an iterable.

This means that, even if we define e.g. signing_keys to be a tuple,
a dev is likely to write:

  Repository(signing_keys=['foo', 'bar'], ...)

And we can make that "work" by adding a conversion to tuple, but
unfortunately a tuple is not equal to a list of same length and
contents, which results in unexpected behavior:

  # assertion fails!
  assert Repository(signing_keys=['key']).signing_keys == ['key']

It seems that if we continue using tuples, users of these classes
would have to be constantly vigilant to remember whether a value
is to be passed and compared as a tuple or list, which is quite
user-unfriendly.

So, to keep the ease-of-use of lists while avoiding the downsides
of tuples, let's define our own immutable list type, FrozenList.
This type is just like a list except that no mutation is allowed,
and instances can be hashed (like a tuple).

This class should be used on attributes like:

- use default=Factory(FrozenList) if you want to default to an empty
  list

- use type=list, and not FrozenList, because `type` acts as the
  documented type and we are not committing to using this specific type

- use converter=FrozenList to convert from input iterables into a
  FrozenList

The resulting behavior is as follows:

    # can create with a plain list
    repo = Repository(signing_keys=['key'])

    # assertion passes
    assert repo.signing_keys == ['key']

    # attempting to modify crashes
    repo.signing_keys.append('other-key')