tests: test invariants on all public model classes
It turns out there are a few mistakes easy to make when defining
new model classes (and we've made them on almost all classes
defined so far). These mistakes can break expected usage of hash/==
and a few other things.

Before fixing this, let's add some tests to cover the problems.
These tests will verify that model objects satisfy basic invariants
such as "actual types match declared types" and hash/==
functionality.

As can be seen in the test, every model class right now
is unable to pass the tests and so are marked as XFAIL.