Ensure Matcher.regex raises when passed non-string
Matcher.regex is documented as accepting a string, but previously
if you passed an already-compiled regex, that would also be
accepted by the validator. This happened because re.compile simply
returns already-compiled regexes without an error. It would
ultimately lead to an error when executing a search, because a
compiled regex is not serializable to JSON.

Fix it to more strictly typecheck the inputs when creating the
matcher, so we fail early and with a meaningful message.