Include task tags in completed/failed logs
Although we were logging when a task completes/fails, our log
messages included only the ID of the completed task. When tasks
were created as part of a larger workflow, this meant that there
was no way to tell which task was doing what, which is generally
useful information. Thus the user of the library would be forced
to add additional logging code.

Let's fix this by including task tags in the logs.  These contain
basic info regarding what a task has done
(e.g. ["pulp:repository:my-repo", "pulp:action:import_upload"]),
making it unnecessary for the caller to explicitly log similar info.