Adjust return type of search_content
This method was returning a list.  It should instead return a Future[Page]:

- because returning a list prevents the user from using a non-blocking code
  style, as otherwise supported by the library

- because returning a list forces eagerly loading all data to be returned
  by the search, which can be unreasonably slow and memory hungry (e.g.
  all-rpm-content repo can have ~300,000 RPMs)

- to be consistent with other methods in the API, e.g. search_repository

Change it. The relevant example is also revised slightly to avoid eagerly
loading all data; it can now log search results page by page.

Note that in the related *_content properties (e.g. repo.rpm_content),
the implementation was adjusted so that these still return a list.
It might be changed in a separate commit.