Fix wrongly mandatory fields for modulemd_defaults
If you look at the schema defined in Pulp, 'stream' and 'profiles' are
not actually mandatory (not part of the unit key).

Although I believe in practice they're expected to be defined for all
real data, the fact that they're not part of the unit key means they're
omitted from responses in some contexts (e.g. units_succcessful in
associate/unassociate calls), therefore they must not be mandatory here
or crashes will occur when handling those responses.

Declaration of 'profiles' was also tweaked for consistency with other
fields: in other classes, a value which is allowed to be None still has
a type declared (as all types are understood to be nullable), so the
fact that this field can be None is not a reason to omit a type here
either.

One test was tweaked to ensure no crashes when stream is missing.

Relates to RHELDST-861.