Make Client support context manager protocol
Clients use thread-based executors under the hood, but we did not
provide any mechanism to shut them down. This is bad design as it means
there's no reliable way to limit the number of spawned threads if you
need to create many clients (e.g. while running a test suite where
new clients are being created in each test).

Make it work in a 'with' statement so it's possible to shut down all
threads associated with the client; attempting to use a client after
shutdown will raise an error. It's not mandatory, but usage of 'with'
is recommended, so several examples and docs were updated to do that.

FakeClient doesn't use any threads but is intended to have as close as
possible behavior to a real Client, so it also raises the same errors
if used after shutdown.

This is similar in spirit to
https://github.com/release-engineering/pushsource/commit/1ebda2c31657d86c68b14c9ea57475976d3e4a07.