Leverage Criteria, search_for_criteria, and PulpSearch
If `type_ids` keyword arg is passed in to `Repository#remove_content` then wrap the list of content type IDs in a `Criteria` object using `_content_type_id` as the field name. This way, when `search_for_criteria` is called, a `PulpSearch` object with its `type_ids` field properly filled out is returned. This can be used to fill out the JSON body of the `unassociate` POST request.