Map RpmUnit.sha256sum to Pulp 'checksum', not 'checksums.sha256'
For RPM units in Pulp, the SHA256 checksum is stored in two places: once
in a 'checksums' dict and once again in a top-level 'checksum' field.

We previously used the former, but only the latter is indexed (being a
part of the unit key). Switching to 'checksum' can make searches much
faster (and can make the difference between a search being feasible at
all or being unusably slow), so let's do that.

This should be safe, as I have confirmed in the Pulp code (linked
within) that the checksum field is always forced to sha256, and also
verified on our live systems that there are no units with a
checksumtype!='sha256'.

Motivation: noticed some queries were slower than expected while working
on RHELDST-5435.