Ensure repository_memberships is set by search_content
Units have a repository_memberships field, but it never ends up
populated if we don't set include_repos during our search request.
Let's start populating it. No user-facing option is provided for this as
the performance impact of returning this info is expected to be
negligible.

It can be argued that this was a bug because the fake client already did
return repository_memberships in units, so this looks like an
unintentional behavior difference between real and fake clients.

Needed for push workflow in RHELDST-5435.