Enforce a stable order on repository_memberships
These unit objects are meant to be a canonical representation of a
unit's current state which can be hashed, compared with ==, etc.

This is undermined somewhat if repository_memberships has no ordering
applied: the Pulp server can theoretically return this list in a
different order for multiple searches on the same unit.

Let's make the field sorted automatically on construction so that we get
more stable behavior in all cases.