Introduce Criteria.with_unit_type
It was already possible for callers to search for units of a specific
type by doing, for example:

  Criteria.with_field("content_type_id", Matcher.in_(["rpm", "srpm"]))

The problem with this is that it requires the caller to know which
values of content_type_id map to each model class such as RpmUnit.
The library already knows this implementation detail internally and it's
best to avoid duplicating it also in the user of the library.

This commit adds a helper which accepts Unit classes as input, as in:

  Criteria.with_unit_type(RpmUnit)

This allows the mapping between unit subclasses and content types to
remain fully encapsulated within this library.