Fix search_content doing repeated blocking queries to /type/
The search_content method wants to know the server's supported content
types so that it can validate requested type IDs, or set defaults if the
caller didn't request any specific type.

The way this was implemented had a couple of related minor issues:

- there was no caching of types, so every call to search_content would
  have to do an additional call to /types/ ; this means code doing
  repeated search_content calls ends up doing 1.5..2x as many HTTP
  requests as needed (depending on the content type).

- the handling of type_ids was written in a blocking style, which is
  inconsistent with the rest of the API aiming to support a non-blocking
  style with all operations being asynchronous.

This commit fixes the above issues for a modest performance improvement
in some cases.