Refactor/fix handling of upload_file in fake client
"iso" units (generic files) are the first, and currently only, type of
content which can be uploaded by this library. There were two issues
with how this was handled on the fake client:

1. uploads were exposed via an `upload_history` property, whose name
   gives the impression of being generic and supporting all kinds of
   uploads, but whose implementation is arguably specific to "iso" units
   as it relies on fields specific to that unit type.

2. uploads were *only* exposed via that property, and would not show up
   in subsequent content searches. This goes against the intent of the
   fake client to behave in a realistic manner (at least with respect to
   CRUD operations). If the fake client is used to upload content into a
   repo, and then search for content within the repo, it should be able
   to find whatever was just uploaded.

Fix (2) so that the primary side-effect of an upload (make unit(s) exist
in a repo) is upheld for the fake client. With that behavior in place,
"fix" (1) by deprecating it as there's no clear reason why we'd need
this separate property if uploaded units can be queried normally.

Motivation: for RHELDST-5435 I need to start adding support for other
content types, but these problems ought to be fixed first.