Introduce YumRepository.upload_rpm
Support uploading & importing RPMs into repositories. The API is
consistent with FileRepository.upload_file and some minor refactoring
allows most of the code to be shared between both.

The main source of complexity here is support in the fake client; as
we want to ensure that uploading RPMs via the fake client gives
realistic behavior, we need to parse the RPM headers using behavior
consistent with a real Pulp server. This requires pulling in some
additional test dependencies.

The upload-files example was also updated to support RPMs.