Add support for uploading yum repo metadata files
Support yum_repo_metadata_file units, which are typically used
for productid certs. Needed for push command in pubtools-pulp.

Although the unit itself is fairly trivial, some refactoring of the fake
client was needed here in order to ensure we have the same semantics as
a real Pulp server (e.g. uploading a productid twice only ends up with
one file in the repo, rather than two).

This refactoring consists of making the fake client store units in a
similar manner to a real pulp server: repos don't contain units, instead
they contain unit keys, and each unit key references a single unit.
Storing the data in a realistic manner like this helps to ensure that
the fake can't create situations which would be impossible on a real
server.