Add support for uploading modulemd streams
Support uploading a YAML file with modulemd and modulemd-defaults
documents. The upload code is trivial, since Pulp takes care of
parsing the documents.

As usual, most of the complexity is in the fake Pulp client which needs
to reproduce compatible upload behavior as a real Pulp. This uses the
existing modulemd unit classes. Some minor refactoring is needed because
this is the first type of upload which (1) can produce more than one
unit from a single upload, and (2) produces a different unit depending
on which repo is the target of the upload.