Fix flaky test/thread-unsafe usage of requests.Session in TaskPoller
Recently I found that test_publish_cancel was flaky on the py2
test configuration. From reading the code, it seems likely caused by
unsynchronized access to a requests.Session here, as we can receive a
request to cancel a task from one thread while another thread is in the
middle of polling for task updates with the same session.

It's known that requests-mock at least doesn't work correctly with
concurrent access to a single session, and some requests during a test
may end up unmocked. That's exactly what's seen at [1].

Fix it by synchronizing all access to the session. Although this is
often fixed by using a thread-local session, in this case TaskPoller
doesn't own the session or have code for creating a session, so it's
simpler to introduce a lock.

[1] https://github.com/release-engineering/pubtools-pulplib/runs/3820726204