Introduce copy_content for copying between repos
This method provides access to the 'associate' action to get units from
one repo to another. It aims to have consistent behavior with the
existing 'search_content' method with respect to criteria handling; the
difference is that the searched units are put into a target repo, rather
than simply returned.

In order for the FakeClient to provide accurate behavior around
units_successful on completed tasks, it was necessary to add a new
pieces of metadata on fields denoting whether or not they make up part
of the unit_key, as only those fields are exposed after a copy.

A fully functional example script is included which can use the new
method to copy between repos with basic filtering.

Some design choices here are worth pointing out:

 Location of method:
  Copying involves a source and destination repo. The method to request
  a copy could have been added to the source repo, destination repo, or
  no repo at all (the client).

  I decided that the client is the best place, because neither the
  source nor destination repo seems obviously correct (so in practice
  it'd probably be hard to remember when using the API), and also to
  avoid emphasizing either repo as more important than the other in the
  operation (considering that both of them will be locked during the
  copy).

 Repo IDs vs objects:
  The method could have accepted just repo IDs, or it could require
  objects.

  I decided to make it require Repository objects due to the expectation
  that we'll eventually have requirements to change the copy behavior
  based on certain properties of the repo (e.g. "if x is in repo notes,
  then y must be set in override_config"). If there's any chance we need
  full Repository objects to do a copy, it's best to require the caller
  to provide them up-front, because only the caller knows the most
  efficient way to query all repos in scope for their current task.