Fix overly strict YumRepoMetadataFileUnit
The sha256sum field on this model can't be mandatory because it's not a
part of the unit key. Responses from some Pulp APIs such as removing or
copying content will only include unit key fields, so it must be
possible to create a Unit with only those fields.

Fix it to make sha256sum optional and ensure this is covered from a
test.