examples: add dump-content, dumps all content from a pulp server
This new example uses the search_content method to search and log all
units in a Pulp server (optionally filtering to certain types).

This can be used as a kind of stress test while working on changes to
the model in this library. As this script will attempt to parse every
unit in the system, including ancient data of low quality, it can reveal
some loading issues which are otherwise easily missed.