Introduce model for erratum units
This commit adds everything needed in order to search or copy erratum
units in Pulp. The primary motivation is to enable upload of errata
(which is not yet implemented to avoid the commit becoming too large).

Note that the model here is very similar to the ErratumPushItem
model in pushsource library. The main difference is that these classes
need to be more tolerant of bad data / missing fields, since Pulp
itself does not enforce any schema on nested fields of errata (e.g.
pkglist).