Fix too strict 'task' schema
Task data from Pulp comes with a result field, but no specific type is
enforced for the field - different plugins and task types will produce
different types of result.

In the case of a "clear orphans" task, the result would be an integer
(presumably the number of orphans deleted). That would cause a crash on
the schema and the underlying code which assumed a truthy value would
always be a dict.

Make the schema and related code more tolerant of other types so that
tasks of this form can be loaded.