Support upload of comps.xml
Introduce YumRepository.upload_comps_xml, which does:

- parse a given comps.xml file into pulp unit representation
- remove existing comps-related units from target repo
- upload new units

There are a couple of design decisions to be explained:

=== Server vs client-side XML parsing ===

The approach used here is compatible with the behavior used in our
toolchain for the past several years, which is to parse comps.xml
client-side and then upload individual units.

However, Pulp does actually support server-side parsing of comps.xml
since early 2016. Support for this was added at "our" request, but
changes were never made to the tooling to start using that feature.

I did investigate moving to server-side parsing of comps.xml but decided
against it because:

- By inspection I was able to find two bugs which would have to be fixed
  before using it. As the feature may have been unused in the years
  since it was implemented, I feel there is a risk of more issues.

- Since the process can't be handled atomically (even if Pulp is doing
  the parsing), we should validate the XML before upload, otherwise an
  attempt to upload an invalid comps.xml could leave the repo in a
  corrupt state. Client-side parsing into Pulp units ensure we only
  proceed with any writes on Pulp if we're confident we've got data that
  Pulp will be able to handle.

=== Lack of model units ===

Unlike with most other data we can upload, no corresponding *Unit
classes were added for the comps-related units (e.g. PackageGroupUnit).
This means that comps.xml data is effectively write-only for this
library right now.

This was done because there is no specific reason to add them and some
potential downsides to doing so. The publishing tools right now treat
comps.xml as an opaque blob; nothing other than the Pulp upload code
looks "inside" of comps.xml, or looks at Pulp units and makes decisions
based on their content.

This seems like a good thing, as it ensures all handling of comps.xml
remains simple; let's not make the Pulp units too easy to access, so
we can hopefully keep it this way.

If the units turn out to be needed, they can be added later with no
backwards-incompatible changes required.