fake: fix an inaccuracy when replacing iso/file units
The Pulp fake stores units indirectly by key, like a real server, so it
is broadly compatible with real Pulp when it comes to mutating or
replacing units.

However, some unit types have additional behavior on top of the unit key
constraints. One of these is for 'iso' units: the unit key technically
allows multiple units with the same name to be stored in a repo, but
Pulp adds extra code to remove these duplicates. Reproduce the same
logic in our fake for more accurate behavior.