Introduce Client.search_task to search for Pulp tasks
Client has a new search_task api to search for the pulp tasks.
This returns Task objects matching the criteria.

Added the same api to the fake client as well which returns the
tasks added to the controller that match the criteria.
Modified the match_field_eq i.e. EqMatcher for the fake client
to match the values in the list type fields too. It now matches
if one of the values in the list type field is present.