Support setting mutable (pulp_user_metadata) fields
We need to be able to support setting these fields on files and RPMs.
It is necessary to support setting them both during upload and at other
times, for example:

- cdn_path: this must be set before the unit can be published by
  cdn_distributor.

- cdn_published: this must be set only *after* the unit has been
  published to the CDN.

- description: on iso/file units, it should be possible to change the
  value for existing content without having to re-upload it.

This commit adds support for setting any fields under pulp_user_metadata
either during upload or as a standalone operation. Fields which can be
updated in this way are called out in the docs.