Add unit_id field onto unit models
Every unit in Pulp has a unique ID. This was not previously exposed as
it seems a good idea to hide that implementation detail if practical.

Add it now to serve the use-case of adjusting pulp_user_metadata fields
on units, for which the API is:

  /pulp/api/v2/content/units/<content_type>/<unit_id>/pulp_user_metadata/

There is no reasonable way to support this API without including unit_id
on the model, so it must be added.

Most of the complexity here is in the fake client, which must generate
unique IDs for units in a deterministic manner so that tests have stable
behavior.