Set ud_file_release_mappings_2 during repo publish
This is a repo-level note which needs to be set in certain cases to
control how files are displayed in UD.

The design around this note is a bit of a mess:

- the data structure is arbitrary

- data is encoded as JSON even though it could have been stored as
  a real dict

- the fields in this note could just as well be stored on individual
  units, as with the rest of per-file metadata

- the data doesn't get copied between repos when units are copied

- in practice, the release pipeline has never fully supported setting
  this note correctly, so releng are using other low-level tools after
  push to manipulate the values

For all these reasons, I don't want to expose any public API for this
and "taint" both this library and pubtools-pulp. For instance I don't
want to add this note onto the Repository model.

Instead, this behavior has been added as a semi-hidden implementation
detail which occurs whenever a FileRepository is published. The way it
works now is that the push code must store the relevant fields
directly on each unit (as it should have been in the first place);
then the code added here will look at those fields and "compile" them
into the required format on the mapping note.

The code is written so that it won't remove anything from the note, or
touch anything it doesn't understand. Therefore, workflows which use
other tools to manipulate the note should continue working.

This can be considered as something which only exists for compatibility
reasons, as it's now possible to access these unit fields in the same
way as any others rather than as a repo note. Hopefully it can be
removed some day.