refactor: use a namedtuple to improve debuggability
This is a tiny change replacing one usage of a bare tuple with
namedtuple. It provides a modest improvement in debuggability of the
code, due to the fact that __repr__ on Future will log the type of its
result (but not the value).

In other words, with this change some instances of the following in
tracebacks:

    <Future at 0x7f095c170990 state=finished returned tuple>

Will be replaced with the slightly more detailed:

    <Future at 0x7f095c170990 state=finished returned UploadResult>

This can help figure out which part of the code is at fault when a crash
occurs.