Fix errors uploading large files [RHELDST-10124]
For both real and fake clients, the uploading of files to Pulp was
implemented using a recursive style (with the real client using
recursive/chained futures while the fake client used recursive function
calls).

In both cases, this can blow the stack for very large files. I
encountered this while testing upload of a 13GB file. Given the default
sys.recursionlimit of 1000 and pulplib chunk size of 1MB, crashes would
tend to occur for files of 1GB and larger.

Fix it by rewriting the upload functions for both clients to use a plain
loop so there is no longer any deep recursion going on during uploads.
In the real client, since we are trying to be as non-blocking as
possible, a new executor was introduced dedicated to the upload
function, and we still support uploading multiple chunks at once.

The default chunk size has also been tuned to a more reasonable value,
as the previous size of 1MB seems a bit small. 10MB is the value which
has actually been used in practice historically.