Improve verbosity of upload logs
While creating/deleting upload IDs, let's log the name of the related
file. While uploading chunks, let's log progress info every now and
then.

The motivation for improving the logs is that there still seem to be
some performance issues with uploading large files, and it's hard to
tell what's going on without this info.