Added filters to remove content from a repository
Repository.remove_content filters content only on type_ids.
This patch adds the option to provide `criteria` which will
filter the contents for removal based on the criteria provided.
It should be the default way to filter contents and type_ids
is deprtecated and will be removed in future version.