Make most classes slots-based
Some tasks using this library are having memory usage issues. Slotted
classes should reduce memory usage, so let's use them where we can.

Note that this doesn't target *all* attr.s classes in the library
because some of them are currently written to make use of self.__dict__,
which doesn't exist for slotted classes. However, all the unit classes,
plus tasks, are covered. These are the important ones since it's typical
for code to be dealing with tens of thousands of these objects, while
only a handful of the other kinds of objects.